refactor testcode + excelfeature test
