+++ src/main/java/innohack/gem/filegen/CsvFileGenerator.java
@@ -99,13 +99,13 @@ public class CsvFileGenerator {
   }
 
 
-  public static void generateFixedCustomerCsvFiles(int numOfFiles, Path dest, int custVariable)
+  public static void generateFixedCustomerCsvFiles(int numOfFiles, Path dest, int custVariable,
+      String filename)
       throws FileNotFoundException {
 
-    String filenamePrefix = "customer_";
+
     for (int i = 0; i < numOfFiles; i++) {
-      String filename = filenamePrefix + i + ".csv";
-      GenUtil.writeToFile(
+         GenUtil.writeToFile(
           generateFixedDataTables(100, custVariable).stream()
               .map(CsvFileGenerator::toCsv)
               .map(row -> String.join(",", row))
